fix(trainer): fix test failures in collect_data and train_pipeline modules
- Fixed test_collect_data_from_multiple_snapshots by adding base_snapshot_id parameter
  to prevent snapshot ID collisions between files
- Fixed test_load_dataset_with_validation by adding processes_per_snapshot parameter
  to create enough processes to pass validation requirements
- Updated all related tests in test_collect_data.py and test_train_pipeline.py
- All tests now pass except for ONNX export limitation (CatBoost issue)

ST-668, ST-669

+++ smoothtask-core/src/metrics/process.rs
@@ -1899,7 +1899,7 @@ mod cache_tests {
         // чтобы гарантировать разные временные метки
         for i in 0..5 {
             let record = ProcessRecord {
-                pid: i as i32,
+                pid: i,
                 ppid: 0,
                 uid: 1000,
                 gid: 1000,
@@ -1958,7 +1958,7 @@ mod cache_tests {
         // Добавляем несколько записей
         for i in 0..5 {
             let record = ProcessRecord {
-                pid: i as i32,
+                pid: i,
                 ppid: 0,
                 uid: 1000,
                 gid: 1000,
